Update Day Template
