+++ 2023/day-template/src/bin/dp2.rs
@@ -1,4 +1,4 @@
-use day::{part2::process, Error};
+use day::{error::Error, part2::process};
 
 #[tracing::instrument]
 fn main() -> Result<(), Error> {
