+++ 2023/day-template/src/lib.rs
@@ -1,23 +1,3 @@
-use std::fmt;
-
+pub mod error;
 pub mod part1;
 pub mod part2;
-
-#[derive(Debug, thiserror::Error)]
-pub enum Error {
-    NotYetImplemented,
-    Other(String),
-}
-impl From<&str> for Error {
-    fn from(value: &str) -> Self {
-        Self::Other(value.to_string())
-    }
-}
-impl fmt::Display for Error {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        match self {
-            Self::Other(msg) => write!(f, "Error: {}", msg),
-            _ => write!(f, "{:?}", self),
-        }
-    }
-}
