+++ 2023/day-template/src/part2.rs
@@ -1,4 +1,4 @@
-use crate::Error;
+use crate::error::Error;
 
 pub fn process(_input: &str) -> miette::Result<String, Error> {
     Err(Error::NotYetImplemented)
